Modified our app to handle Post requests.
The server will listen to "data" and "end" events which will assemble the post body and call the router will the postbody as argument respectively. Router passes the post data to the requesthandler, which can act on it or display it.

By listening to "data" and "end" we handle postrequest asynchronously, which is better in case of large data is posted.

pg. 46