Completed file upload, save to disk and show functionality.
Uses the formidable module which parses the request. Therefore, the request object needs to get passed from the server to the requesthandler via the router (we want the use formidable in the requesthanler). As a consequence, passing postData is no longer needed (nor possible for correct functioning).

The upload requesthandler uses formidble to parses the request and rename the form to /tmp/test.pgn. It uses the 'show' uri to let the show requesthandler display the png file.

We have our working app!

A bit strange to pass our entire request to a requesthandler though.

pg. 58